+++ src/webapp/controllers/UsersController.php
@@ -121,7 +121,7 @@ class UsersController extends Controller
 
     public function destroy($username)
     {
-        if($this->auth->isAdmin() === true){
+        if($this->auth->check() && $this->auth->isAdmin() === true){
             if ($this->userRepository->deleteByUsername($username) === 1) {
                 $this->app->flash('info', "Sucessfully deleted '$username'");
                 $this->app->redirect('/admin');
